+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added DF and SL Blacksmithing recipes
 - Added DF and SL Enchanting recipes
 - Added DF and SL Engineering recipes
+- Added DF and SL Inscription recipes
 ### Changed
 - Significant additions, fixes and changes to Vanilla instance loot